requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
httpx
requests

# LangChain & Graph Data structures
langchain
langchain-community
langchain-huggingface
langchain-groq
langgraph
langgraph-checkpoint-sqlite

# Data Parsers & Vectors
PyMuPDF
chromadb

# Local Testing Generators
fpdf2
